feat(core): changed from UnsignedByte to traditional byte use.
- Modified UnsignedByte class to be a static class that does the necessary conversions to use unsigned bytes.

+++ src/main/java/com/acousea/backend/core/communicationSystem/domain/communication/constants/RoutingChunk.java
@@ -57,8 +57,8 @@ public final class RoutingChunk {
     }
 
     public String encode() {
-        return Integer.toHexString(sender.getValue().toInt())
-                + Integer.toHexString(receiver.getValue().toInt())
+        return Integer.toHexString(sender.getValue())
+                + Integer.toHexString(receiver.getValue())
                 + Integer.toHexString(TTL);
     }
